Deprecates "composeMocks" in favor of "setupWorker"
