+++ src/composeMocks/setupWorker.test.ts
@@ -1,4 +1,4 @@
-import { composeMocks } from './composeMocks'
+import { setupWorker } from './setupWorker'
 import rest, { restContext } from '../handlers/rest'
 import { MockedRequest, ResponseResolver } from '../handlers/requestHandler'
 
@@ -11,7 +11,7 @@ test('Generates schema based on provided handlers', () => {
     return res(json({ a: 2 }))
   }
 
-  const api = composeMocks(
+  const api = setupWorker(
     rest.get('https://api.github.com/users/:username', simpleResolver),
     rest.get('foo', simpleResolver),
     rest.post(/footer/, simpleResolver),
