+++ test/graphql-api/mutation.mocks.ts
@@ -1,6 +1,6 @@
-import { composeMocks, graphql } from 'msw'
+import { setupWorker, graphql } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   graphql.mutation('Logout', (req, res, ctx) => {
     return res(
       ctx.data({
@@ -12,4 +12,4 @@ const { start } = composeMocks(
   }),
 )
 
-start()
+worker.start()
