+++ test/msw-api/exception-handling.mocks.ts
@@ -1,6 +1,6 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('https://api.github.com/users/:username', () => {
     // @ts-ignore
     nonExisting()
@@ -8,4 +8,4 @@ const { start } = composeMocks(
   }),
 )
 
-start()
+worker.start()
