+++ test/msw-api/integrity-check-invalid.mocks.ts
@@ -1,12 +1,12 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('https://api.github.com/users/:username', (req, res, ctx) => {
     return res(ctx.json({ mocked: true }))
   }),
 )
 
-start({
+worker.start({
   serviceWorker: {
     // Use custom Service Worker URL for the purpose of intentionally
     // registering an outdated worker. Please do not use this as an example.
