+++ test/msw-api/setup-worker/start/start.mocks.ts
@@ -1,13 +1,13 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('/user', (req, res, ctx) => {
     return res(ctx.status(200))
   }),
 )
 
 // @ts-ignore
-window.__MSW_REGISTRATION__ = start().then((reg) => {
+window.__MSW_REGISTRATION__ = worker.start().then((reg) => {
   console.log('Registration Promise resolved')
   return reg.constructor.name
 })
