+++ test/msw-api/unregister.mocks.ts
@@ -1,6 +1,6 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('https://api.github.com', (req, res, ctx) => {
     return res(
       ctx.json({
@@ -11,4 +11,4 @@ const { start } = composeMocks(
 )
 
 // @ts-ignore
-window.__mswStart = start
+window.__mswStart = worker.start
