+++ test/rest-api/headers-multiple.mocks.ts
@@ -1,6 +1,6 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.post('https://test.msw.io/', (req, res, ctx) => {
     return res(
       ctx.json({
@@ -21,4 +21,4 @@ const { start } = composeMocks(
   }),
 )
 
-start()
+worker.start()
