+++ test/rest-api/request-matching/uri.mocks.ts
@@ -1,6 +1,6 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('https://api.github.com/made-up', (req, res, ctx) => {
     return res(
       ctx.json({
@@ -28,4 +28,4 @@ const { start } = composeMocks(
   }),
 )
 
-start()
+worker.start()
