+++ test/rest-api/xhr.mocks.ts
@@ -1,9 +1,9 @@
-import { composeMocks, rest } from 'msw'
+import { setupWorker, rest } from 'msw'
 
-const { start } = composeMocks(
+const worker = setupWorker(
   rest.get('https://api.github.com/users/octocat', (req, res, ctx) => {
     return res(ctx.json({ mocked: true }))
   }),
 )
 
-start()
+worker.start()
